# titan-web-client/src/http/api/openapi/titan.v1.yaml
openapi: 3.0.2
info:
  title: Titan
  version: '1.0'
  license:
    name: MIT
    url: 'https://github.com/unkso/titan-app/blob/master/LICENSE'
servers:
  - url: 'http://localhost:3000'
paths:
...
  '/api/organizations/{org_id}/roles:reorder':
    parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: reorder_roles
      tags: []
      responses:
        '200':
          description: OK
      operationId: 'post-organizations-org_id-roles:reorder'
      security:
        - api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: reorder_organization_roles_fields
              type: object
              properties:
                role_ids:
                  type: array
                  items:
                    type: integer
...
components:
...
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header